tests-regtest/pyproject.toml
[project]
name = "hold"
version = "0.1.0"
license = "MIT"
description = "Tests for CLN hold invoice plugin"
requires-python = ">=3.10"
dependencies = [
    "bolt11>=2.1.1",
    "grpcio>=1.73.1",
    "grpcio-tools>=1.73.1",
    "pytest>=8.4.1",
    "ruff>=0.12.1",
]

[tool.ruff]
exclude = ["hold/protos"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D107",
    "D211",
    "D212",
    "S605",
    "D203",
    "ISC001",
    "COM812",
    "S101",
    "PLR2004",
    "PT011",
    "TC001",
    "F403",
    "F405",
]
